split bulk request in chunks
