+++ src/platform/packages/private/kbn-index-editor/src/services/rows_virtual_indexes.test.ts
